Make the file upload work.
The change in the data service removes the Content-Type header and
lets the XMLHttpRequest set the "multipart/form-data" value with
the appropriate value that contains the boundary information.

The change in addimage.component.ts changes the name of the
submitted field from 'uploadFile' to 'file' because the Java back-
end expects the field to be named 'file'.